requirements.txt
streamlit
torch
faiss-cpu
scikit-learn
transformers